Cleaning up project
Meaningful name, yeah, but all I`ve done in this commit
is that I added some DTo classes, added save methods to every service
and added analogue controller endpoints.

+++ src/main/java/com/it/cinemabackend/model/dto/ShowtimeDTO.java
@@ -13,8 +13,8 @@ import lombok.ToString;
 @ToString
 public class ShowtimeDTO {
 
-    private Integer id;
-    private Integer movieId;
+    private Long id;
+    private Long movieId;
     private String title;
     private List<String> genres;
     private String ageRestriction;
